add check user system prevous remove dictionary

+++ server/controllers/dictionaryController.js
@@ -63,7 +63,6 @@ exports.getOneDictionary = async (req, res) => {
 
 exports.editDictionary = async (req, res) => {
     const { id, newvocabulary } = req.params;
-    console.log(id, JSON.parse(newvocabulary))
     try {
         if (id) {
             dictionaryModel.updateOne({ _id: id }, { vocabulary: JSON.parse(newvocabulary) }, (err) => {
@@ -83,16 +82,21 @@ exports.editDictionary = async (req, res) => {
 }
 
 exports.deleteDictionary = async (req, res) => {
-    const { id } = req.params;
-    console.log(id)
+    const { id, userid, login } = req.params;
     try {
-        if (id) {
-            dictionaryModel.deleteOne({ _id: id }, (err) => {
-                if (err)
-                    return console.log(err)
+        if (id && userid && login) {
+            const findUser = await userModel.find({ _id: userid, login: login });
+            if (findUser.length > 0) {
+                dictionaryModel.deleteOne({ _id: id }, (err) => {
+                    if (err)
+                        return console.log(err)
 
-                res.status(200).json({ status: 'correct' });
-            })
+                    res.status(200).json({ status: 'correct' });
+                })
+            }
+            else {
+                res.status(500).json({ message: 'error' });
+            }
         }
         else {
             res.status(200).json({ status: 'incorrect' });
